+++ crates/gpui/src/platform/app_identifier.rs
@@ -1,6 +1,13 @@
 use std::sync::OnceLock;
 
-use windows::Win32::Foundation::MAX_PATH;
+use util::ResultExt;
+use windows::Win32::{
+    Foundation::{CloseHandle, HANDLE, MAX_PATH},
+    System::Memory::{
+        MapViewOfFile, OpenFileMappingW, UnmapViewOfFile, FILE_MAP_READ, FILE_MAP_WRITE,
+    },
+};
+use windows_core::HSTRING;
 
 static APP_IDENTIFIER: OnceLock<String> = OnceLock::new();
 static APP_MUTEX_IDENTIFIER: OnceLock<String> = OnceLock::new();
@@ -61,4 +68,49 @@ pub fn get_app_shared_memory_identifier() -> &'static str {
         }
         identifier
     })
+}
+
+/// TODO:
+pub fn read_dock_action_argument(shared_memory_handle: HANDLE) -> String {
+    unsafe {
+        let memory_addr = MapViewOfFile(shared_memory_handle, FILE_MAP_READ, 0, 0, 0);
+        let string = String::from_utf8_lossy(std::slice::from_raw_parts(
+            memory_addr.Value as *const _ as _,
+            APP_SHARED_MEMORY_MAX_SIZE,
+        ))
+        .trim_matches('\0')
+        .to_string();
+        UnmapViewOfFile(memory_addr).log_err();
+        string
+    }
+}
+
+/// TODO:
+pub fn write_dock_action_argument(message: &str) {
+    if message.len() > APP_SHARED_MEMORY_MAX_SIZE {
+        log::error!(
+            "The length of the message to send should be less than {APP_SHARED_MEMORY_MAX_SIZE}"
+        );
+        return;
+    }
+    unsafe {
+        let msg = message.as_bytes();
+        let pipe = OpenFileMappingW(
+            FILE_MAP_WRITE.0,
+            false,
+            &HSTRING::from(get_app_shared_memory_identifier()),
+        )
+        .unwrap();
+        let memory_addr = MapViewOfFile(pipe, FILE_MAP_WRITE, 0, 0, 0);
+        // Clear the buffer first
+        let empty_buffer = vec![0u8; APP_SHARED_MEMORY_MAX_SIZE];
+        std::ptr::copy_nonoverlapping(
+            empty_buffer.as_ptr(),
+            memory_addr.Value as _,
+            empty_buffer.len(),
+        );
+        std::ptr::copy_nonoverlapping(msg.as_ptr(), memory_addr.Value as _, msg.len());
+        UnmapViewOfFile(memory_addr).log_err();
+        CloseHandle(pipe).log_err();
+    }
 }
