+++ crates/zed/src/main.rs
@@ -244,7 +244,7 @@ fn main() {
     if failed_single_instance_check {
         println!("zed is already running");
         #[cfg(target_os = "windows")]
-        if let Some(ref argument) = args.new_instance {
+        if let Some(ref argument) = args.dock_action {
             zed::windows_only_instance::send_instance_message(argument);
         }
         return;
